break up ideas class
