+++ controller/base.php
@@ -13,7 +13,8 @@ namespace phpbb\ideas\controller;
 use phpbb\auth\auth;
 use phpbb\config\config;
 use phpbb\controller\helper;
-use phpbb\ideas\factory\ideas;
+use phpbb\ideas\ext;
+use phpbb\ideas\factory\manager as ideas;
 use phpbb\ideas\factory\linkhelper;
 use phpbb\language\language;
 use phpbb\pagination;
@@ -144,7 +145,7 @@ abstract class base
 			'S_SEARCHBOX_ACTION'	=> append_sid("{$this->root_path}search.{$this->php_ext}"),
 			'S_SEARCH_IDEAS_HIDDEN_FIELDS'	=> build_hidden_fields(['fid' => [$this->config['ideas_forum_id']]]),
 
-			'U_SEARCH_MY_IDEAS' 	=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ideas::SORT_MYIDEAS, 'status' => '-1']),
+			'U_SEARCH_MY_IDEAS' 	=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ext::SORT_MYIDEAS, 'status' => '-1']),
 		]);
 	}
 }
