+++ controller/index_controller.php
@@ -11,12 +11,10 @@
 namespace phpbb\ideas\controller;
 
 use phpbb\exception\http_exception;
-use phpbb\ideas\factory\ideas;
+use phpbb\ideas\ext;
 
 class index_controller extends base
 {
-	const NUM_IDEAS = 5;
-
 	/**
 	 * Controller for /ideas
 	 *
@@ -31,21 +29,21 @@ class index_controller extends base
 		}
 
 		// Generate latest ideas
-		$ideas = $this->ideas->get_ideas(self::NUM_IDEAS, ideas::SORT_DATE, 'DESC');
+		$ideas = $this->ideas->get_ideas(ext::NUM_IDEAS, ext::SORT_DATE, 'DESC');
 		$this->assign_template_block_vars('latest_ideas', $ideas);
 
 		// Generate top ideas
-		$ideas = $this->ideas->get_ideas(self::NUM_IDEAS, ideas::SORT_TOP, 'DESC');
+		$ideas = $this->ideas->get_ideas(ext::NUM_IDEAS, ext::SORT_TOP, 'DESC');
 		$this->assign_template_block_vars('top_ideas', $ideas);
 
 		// Generate recently implemented
-		$ideas = $this->ideas->get_ideas(self::NUM_IDEAS, ideas::SORT_DATE, 'DESC', ideas::$statuses['IMPLEMENTED']);
+		$ideas = $this->ideas->get_ideas(ext::NUM_IDEAS, ext::SORT_DATE, 'DESC', ext::$statuses['IMPLEMENTED']);
 		$this->assign_template_block_vars('implemented_ideas', $ideas);
 
 		$this->template->assign_vars(array(
-			'U_VIEW_TOP'		=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ideas::SORT_TOP]),
-			'U_VIEW_LATEST'		=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ideas::SORT_NEW]),
-			'U_VIEW_IMPLEMENTED'=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ideas::SORT_DATE, 'status' => ideas::$statuses['IMPLEMENTED']]),
+			'U_VIEW_TOP'		=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ext::SORT_TOP]),
+			'U_VIEW_LATEST'		=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ext::SORT_NEW]),
+			'U_VIEW_IMPLEMENTED'=> $this->helper->route('phpbb_ideas_list_controller', ['sort' => ext::SORT_DATE, 'status' => ext::$statuses['IMPLEMENTED']]),
 			'U_POST_ACTION'		=> $this->helper->route('phpbb_ideas_post_controller'),
 			'U_MCP' 			=> ($this->auth->acl_get('m_', $this->config['ideas_forum_id'])) ? append_sid("{$this->root_path}mcp.{$this->php_ext}", "f={$this->config['ideas_forum_id']}&amp;i=main&amp;mode=forum_view", true, $this->user->session_id) : '',
